+++ prototipo/app/Http/Controllers/Evaluacion.php
@@ -6,19 +6,33 @@ use Illuminate\Http\Request;
 use App\Models\EvaluacionModel;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Http\RedirectResponse;
+use Illuminate\Validation\ValidationException; 
 
 class Evaluacion extends Controller
 {
 
     public function index()
     {
+        if (!session()->has('user')) {
+            return redirect()->route('login');
+        } 
+
         $evaluaciones = EvaluacionModel::paginate(10);
         return view('administrador/evaluacion/index', compact('evaluaciones'));
     }
 
 
-    public function busqueda()
+    public function busqueda(Request $request)
     {
+        if (!session()->has('user')) {
+            return redirect()->route('login');
+        } 
+
+        $request->validate([
+            'valorBusqueda' => 'required|string',
+        ]);
+
         $evaluaciones = DB::table('evaluacion')
             ->where('actividadAprendizaje', 'like', '%' . $_GET['valorBusqueda'] . '%')
             ->orwhere('tipoEvaluacion', 'like', '%' . $_GET['valorBusqueda'] . '%')
@@ -30,12 +44,26 @@ class Evaluacion extends Controller
 
     public function agregar()
     {
+        if (!session()->has('user')) {
+            return redirect()->route('login');
+        } 
+
         return view('administrador/evaluacion/agregar');
     }
 
 
     public function insert(Request $request)
     {
+        if (!session()->has('user')) {
+            return redirect()->route('login');
+        }
+
+        $request->validate([
+            'actividadAprendizaje' => 'required|string',
+            'tipoEvaluacion' => 'required|string',
+            'archivoEjemplo'=> 'required',
+        ]);
+
         $archivo = null;
         if ($request->hasFile('archivoEjemplo')) {
             $archivo = Storage::disk('public')->put('/ejemplos', $request->file('archivoEjemplo'));
@@ -55,6 +83,11 @@ class Evaluacion extends Controller
     public function ver(string $id)
     {
 
+        if (!session()->has('user')) {
+            return redirect()->route('login');
+        }
+        
+
         $evaluaciones = DB::table('evaluacion')
             ->where('actividadAprendizaje', 'like', '%' . $_GET['valorBusqueda'] . '%')
             ->orwhere('tipoEvaluacion', 'like', '%' . $_GET['valorBusqueda'] . '%')
@@ -66,6 +99,10 @@ class Evaluacion extends Controller
 
     public function editar(string $id)
     {
+        if (!session()->has('user')) {
+            return redirect()->route('login');
+        } 
+
         $evaluacion = EvaluacionModel::find($id);
         return view('administrador/evaluacion/editar', compact('evaluacion'));
     }
@@ -73,6 +110,16 @@ class Evaluacion extends Controller
 
     public function update(Request $request, $id)
     {
+        if (!session()->has('user')) {
+            return redirect()->route('login');
+        } 
+
+        $request->validate([
+            'actividadAprendizaje' => 'required|string',
+            'tipoEvaluacion' => 'required|string',
+            'archivoEjemplo'=> 'required',
+        ]);
+
         $archivo = null;
         if ($request->hasFile('archivoEjemplo')) {
             $archivo = Storage::disk('public')->put('/evaluaciones', $request->file('archivoEjemplo'));
@@ -90,6 +137,10 @@ class Evaluacion extends Controller
 
     public function eliminar(string $id)
     {
+        if (!session()->has('user')) {
+            return redirect()->route('login');
+        } 
+        
         $evaluacion = EvaluacionModel::find($id);
         $evaluacion->delete();
